[FE] refactor: 누락된 import 위치 불일치 문제 해결 #59
